backend/pyproject.toml
[project]
name = "homesite-v2"
version = "2.0.0"
description = "Home automation system — API backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    # Validation & settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # MQTT (device gateway)
    "aiomqtt>=2.0.0",
    # HTTP client (backend -> device gateway)
    "httpx>=0.27.0",
    # Logging
    "structlog>=24.0.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # WebSocket
    "websockets>=13.0",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.30.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
